Revert "Send async_hooks to instances, not classes, in rails 3"
Although Rails 3 calls callbacks in instance context rather than class context,
and although we've updated all the callbacks in our app to use instance context,
that really only applies to normal Active Record callbacks.

The special "async" variants of the callbacks (which we use in three models)
still need to be called in the old way (ie. explicitly passing in the record
object). That in turn means that the async changes previously introduced and
reverted need to be reinstated.

We'll be making a follow-up commit in causes.git to bring those 3 models into
line.

This reverts commit bc23f962.

Change-Id: Iede431ca15d44e4dfabfe10112dbb2a137b0b2e1
Reviewed-on: https://gerrit.causes.com/4284
Reviewed-by: Alex Wallace <alex.wallace@causes.com>
Tested-by: Alex Wallace <alex.wallace@causes.com>