Remove set_close_on_exec from MysqlAdapter
This method was throwing warnings for us on production since we are
using the C mysql binding and we apparently can't use this method with
it.

Furthermore, in Rails 3.2 this extension of the AbstractMysqlAdapter was
interfering with the updated activerecord and blowing up as a superclass
mismatch.

Killing two birds with one stone.

Change-Id: Iafd62ee8c87ed38edab582421e5ec03cc1d12e02
Reviewed-on: https://gerrit.causes.com/5252
Reviewed-by: Alex Wallace <alex.wallace@causes.com>
Tested-by: Alex Wallace <alex.wallace@causes.com>