Massively boost the default TTR on all Beanstalk jobs
In recent times we've observed an increase in the number of
Beanstalk::NotFoundError incidences in Airbrake:

  http://causes.airbrakeapp.com/errors/1805629

We've caught specific examples of long-running jobs in the logs which lead us
to believe that we are exceeding the default "TTR" set by the async_observer
plug-in. As described here:

  https://github.com/kr/beanstalkd/blob/master/doc/protocol.txt#L202

this meant that any job running for longer than the TTR (in our case, 120
seconds) would be reclassified as ready. This mean that long-running jobs would
be picked up by another worker and redundantly run. Examples of such jobs are
the donor CSV download, that can take as long as 15 minutes or more.

Once a job has been reserved (or re-reserved) by another worker, attempting to
delete it will yield a Beanstalk::NotFoundError. Our testing suggests that a
job can get into an "infinite" loop, retrying until it ultimately gets buried.

All this is bad, for several reasons:

  - one is that it is a waste of resources, and probably played a large part in
    the total annihilation of our Beanstalk workers this morning
  - another is that if any of these jobs have user-visible side effects, such
    as sending email, then the user experience suffers in tangible ways
  - finally, it generates a scary amount of Airbrake noise which is hard to
    diagnose because it looks like a broken connection to Beanstalk, but is in
    fact a failure to delete an already-deleted job

So, the fix is to boost the TTR to an absurdly large value. We figure the
potential damage of running the same job multiple times is worse than the
alternative of having a slow response to an expired worker. Even the
possibility of losing jobs is probably not worth worrying about compared to the
risk of duplicate work done on jobs.

I guess one takeaway from all this is that we should avoid writing
very-long-running async jobs whenever possible, favoring splitting them into
smaller jobs. Given the "infinite" looping behavior, we really don't want jobs to
ever hit their TTR; more than a safety feature, it looks like a dangerous "run
at all costs" design.

Change-Id: Ie4ae0d5328402a5c01e9ebfbe0e3bde73aaa8d5f
Reviewed-on: https://gerrit.causes.com/6211
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>