Workers shutdown on TERM when queue is empty
Workers were already shutting down on TERM immediately after
processing a job, but would never shutdown while the queue was empty,
instead blocking indefinitely waiting to reserve a job.

This adds a timeout to the beanqueue reserve call and bails out of
the get_job loop if a TERM is sent.

Also changed the TERM handling code to call the finish callback as
the INT handling already did, which seems more correct.

Change-Id: Icf38dd17ab585a88e395e7c1c44cea259dac2f5d
Reviewed-on: https://gerrit.causes.com/11849
Tested-by: Lann Martin <lann@causes.com>
Reviewed-by: Elliot Block <elliot.block@causes.com>
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>