Add job duration logging to async workers
This change makes async workers log the name of the code they were
running and the time duration of the job on the same line.

This allows us to measure how often and how long each type of job
runs, which will hopefully allow us to diagnose what's bottlenecking
our beanqueue.

The log lines are recorded in the following format:

    #!job-duration!#{job_duration_milliseconds}!#{job[:code]}

Release Note:
  (Eng) Async workers now log the name of the code they were running,
  and the duration of time that the job took.

Related story:
  http://www.pivotaltracker.com/story/show/35798063
  ("Bean queue traffic is backed up again")

Change-Id: I719eedbf99c2e515873b39ba0a9f8f962bbdc1f8
Reviewed-on: https://gerrit.causes.com/12128
Reviewed-by: Elliot Block <elliot.block@causes.com>
Tested-by: Elliot Block <elliot.block@causes.com>