Only delete pidfile when worker daemon exits, not child processes
The worker handles removal of its pidfile in an at_exit block. If the
worker forks, its child process would inherit the at_exit block and
end up deleting the pidfile when it died.

This checks Process.pid in the at_exit block to ensure that only the
worker process itself will delete the file.

Change-Id: I42be0a32069886a1a72247cbd86ab6c7d9919f6a
Reviewed-on: https://gerrit.causes.com/12590
Tested-by: Lann Martin <lann@causes.com>
Reviewed-by: Adam Derewecki <adam@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>