Break out of main loop using break
Previously, the workers broke out of the main loop by raising an
Interrupt. This is probably more clever than we need to be.
Additionally, I am pretty sure a TERM signal is not delivered until a
blocking operation, e.g. select(), times out, which in some cases can
take upwards of 30+ seconds. It's even possible the signal is never
delivered to the process.

Implementing a sentinel file should be more reliable/graceful than the
TERM, sleep, TERM, sleep, KILL that monit currently does. Actually, I
lied, monit currently does a TERM, sleep, KILL. I am about to change
it to TERM twice.

We can touch the sentinel file, deploy new code as usual, then issue a
monit command, *guaranteed* to stop the workers, rm the sentinel file,
and issue a monit start command.

This is more reliable because it:
  * Ensures the process shuts down
  * Gives the process more time to gracefully shut down
  * Doesn't require monit to send a stop command to 100+ workers

Change-Id: I9214aaa217340dea417b9564e4e592c93fd86e7d
Reviewed-on: https://gerrit.causes.com/13590
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>