Ensure workers exit gracefully when sentinel file is touched
Touching the sentinel file was indeed causing the workers to shut down,
but not as gracefully as we'd intended.

Breaking inside `#safe_dispatch` was causing the method to return nil,
which in turn made the logging code explode when we did a `job.body`.

The fix is to defer to the outer "main" loop.

Change-Id: Ic0c339eba30a26007c135266780f87d3c2cbf600
Reviewed-on: https://gerrit.causes.com/13648
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>