Remove __FILE__-relative paths from executable
Since these files are no longer included in the Causes repo's
vendor/gems/ directory, we need to remove these relative paths. I chose
this path (which depends on ENV['PWD']) because it is simpler than, say,
adding another command-line option to specify the Rails root. But
long-term, I feel like that is probably the better way to approach this.

Change-Id: I0160c78397581ad2927a137e16ea865d154f9b27
Reviewed-on: http://gerrit.causes.com/44436
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>
Tested-by: Tom Dooner <tom.dooner@brigade.com>