Correctly handle failing Job#[] in ensure block
Right now we have a job enqueued that is the empty string. The (not
maintained by us) beanstalk-client-ruby gem raises an error in this case
when trying to fetch `job[:code]`. Since this is in an `ensure` block,
it is always run, regardless of whether the job was valid or not. If the
ruby process crashes, I believe beanstalk will then re-issue the job to
another worker, which subsequently will also crash.

Change-Id: I177d7d9ef6651b7e07b8cb079d853700cb2dbfce
Reviewed-on: https://code.brigade.zone/7627
Reviewed-by: Tom Dooner <tom.dooner@brigade.com>
Tested-by: Tom Dooner <tom.dooner@brigade.com>